add IN_DanhGiaLKhaoSat page

+++ Controllers/OutPatients/OUT_CauHoiKhaoSatController.cs
@@ -46,7 +46,7 @@ namespace feedBackMvc.Controllers.OutPatients
         {
             public List<string>? TieuDeCauHois { get; set; }
             public List<string>? CauHois { get; set; }
-            public int Id {get; set;}
+            public int Id { get; set; }
         }
         [HttpPost]
         public async Task<IActionResult> ThemCauHoiKhaoSat([FromBody] TitleAndQuestionList data)
@@ -77,7 +77,7 @@ namespace feedBackMvc.Controllers.OutPatients
                     if (existingRecord != null)
                     {
                         await transaction.RollbackAsync();
-                        return Json(new { success = false, message = "Tiêu đề đã tồn tại"});
+                        return Json(new { success = false, message = "Tiêu đề đã tồn tại" });
                         //return Json(new { success = false });
                     }
                     _context.OUT_CauHoiKhaoSat.Add(newGroup);
@@ -90,15 +90,15 @@ namespace feedBackMvc.Controllers.OutPatients
             {
                 await transaction.RollbackAsync();
                 _logger.LogError(ex, "Error occurred while adding records.");
-            return StatusCode(500, $"Internal server error: {ex.Message}");
-             //return Json(new { success = false });
+                return StatusCode(500, $"Internal server error: {ex.Message}");
+                //return Json(new { success = false });
             }
         }
         public class DeleteRequest
         {
             public int Id { get; set; }
         }
-       [HttpPost]
+        [HttpPost]
         public async Task<IActionResult> XoaCauHoiKhaoSat([FromBody] DeleteRequest request)
         {
             try
@@ -108,7 +108,11 @@ namespace feedBackMvc.Controllers.OutPatients
                 {
                     return Json(new { success = false, message = "Nhóm câu hỏi không tồn tại." });
                 }
-
+                bool isUsedInInSurvey = await _context.OUT_MauKhaoSat.AnyAsync(m => m.CauHoiKhaoSat != null && nhom.TieuDeCauHoi != null && m.CauHoiKhaoSat.Contains(nhom.TieuDeCauHoi));
+                if (isUsedInInSurvey)
+                {
+                    return Json(new { success = false, message = "Câu hỏi này đang được sử dụng trong mẫu khảo sát." });
+                }
                 _context.OUT_CauHoiKhaoSat.Remove(nhom);
                 await _context.SaveChangesAsync();
 
@@ -131,26 +135,30 @@ namespace feedBackMvc.Controllers.OutPatients
         public async Task<IActionResult> CapNhatCauHoiKhaoSat([FromBody] List<UpdateRequest> request)
         {
             if (request == null || !request.Any())
-                {
-                    return BadRequest(new { success = false, message = "Không nhận được dữ liệu cập nhật." });
-                }
-                try{
-                    foreach(var item in request){
-                        // Giả sử bạn có một DbContext tên là _context
-                        var existingItem = await _context.OUT_CauHoiKhaoSat.FindAsync(item.Id);
-                        if(existingItem != null)
-                        {
-                            existingItem.TieuDeCauHoi = item.TieuDeCauHoi;
-                            existingItem.CauHoi = item.CauHoi;
-                            // Bạn có thể thêm các logic kiểm tra khác ở đây
-                        }
-                    }
-                    await _context.SaveChangesAsync();
-                    return Json(new { success = true, message = "Cập nhật thành công." });
-                }catch(Exception ex){
-                    // Xử lý ngoại lệ và ghi log nếu cần
-                    return Json(new { success = false, message = "Có lỗi xảy ra: " + ex.Message });
-                }
+            {
+                return BadRequest(new { success = false, message = "Không nhận được dữ liệu cập nhật." });
             }
+            try
+            {
+                foreach (var item in request)
+                {
+                    // Giả sử bạn có một DbContext tên là _context
+                    var existingItem = await _context.OUT_CauHoiKhaoSat.FindAsync(item.Id);
+                    if (existingItem != null)
+                    {
+                        existingItem.TieuDeCauHoi = item.TieuDeCauHoi;
+                        existingItem.CauHoi = item.CauHoi;
+                        // Bạn có thể thêm các logic kiểm tra khác ở đây
+                    }
+                }
+                await _context.SaveChangesAsync();
+                return Json(new { success = true, message = "Cập nhật thành công." });
+            }
+            catch (Exception ex)
+            {
+                // Xử lý ngoại lệ và ghi log nếu cần
+                return Json(new { success = false, message = "Có lỗi xảy ra: " + ex.Message });
+            }
+        }
     }
 }
